+++ skimage/setup.py
@@ -20,6 +20,7 @@ def configuration(parent_package='', top_path=None):
     config.add_subpackage('measure')
     config.add_subpackage('morphology')
     config.add_subpackage('transform')
+    config.add_subpackage('transform_xy')
     config.add_subpackage('util')
     config.add_subpackage('segmentation')
     config.add_subpackage('external')
