+++ skimage/transform/__init__.py
@@ -1,22 +1,24 @@
 from .histogram_matching import match_histograms
-from .hough_transform import (hough_line, hough_line_peaks,
-                              probabilistic_hough_line, hough_circle,
-                              hough_circle_peaks, hough_ellipse)
+from .hough_transform import (hough_line, probabilistic_hough_line,
+                              hough_circle_peaks, hough_line_peaks,
+                              hough_circle, hough_ellipse)
 from .radon_transform import (radon, iradon, iradon_sart,
                               order_angles_golden_ratio)
 from .finite_radon_transform import frt2, ifrt2
 from .integral import integral_image, integrate
-from ._geometric import (estimate_transform,
-                         matrix_transform, EuclideanTransform,
-                         SimilarityTransform, AffineTransform,
-                         ProjectiveTransform, FundamentalMatrixTransform,
-                         EssentialMatrixTransform, PolynomialTransform,
-                         PiecewiseAffineTransform)
-from ._warps import (swirl, resize, rotate, rescale,
-                     downscale_local_mean, warp, warp_coords)
+from ._warps import resize, rescale, downscale_local_mean
 from .pyramids import (pyramid_reduce, pyramid_expand,
                        pyramid_gaussian, pyramid_laplacian)
 from .seam_carving import seam_carve
+
+from ..transform_xy import (warp, swirl, rotate,
+                            estimate_transform,
+                            matrix_transform, EuclideanTransform,
+                            SimilarityTransform, AffineTransform,
+                            ProjectiveTransform, FundamentalMatrixTransform,
+                            EssentialMatrixTransform, PolynomialTransform,
+                            PiecewiseAffineTransform)
+from .._transform_common import warp_coords
 
 
 __all__ = ['match_histograms',
