+++ skimage/transform_xy/__init__.py
@@ -1,0 +1,9 @@
+from .hough_transform import (hough_line, probabilistic_hough_line,
+                              hough_circle_peaks)
+from ._geometric import (estimate_transform,
+                         matrix_transform, EuclideanTransform,
+                         SimilarityTransform, AffineTransform,
+                         ProjectiveTransform, FundamentalMatrixTransform,
+                         EssentialMatrixTransform, PolynomialTransform,
+                         PiecewiseAffineTransform)
+from ._warps import swirl, rotate, warp
